Re-refactor into raster.py (not the source of error), and fix temporary files in test_save of raster and vector
