+++ doc/source/conf.py
@@ -93,8 +93,8 @@ plot_formats = [(".png", 400)]
 
 # To avoid long path names in inheritance diagrams
 inheritance_alias = {
-    "geoutils.raster.core.Raster": "geoutils.Raster",
-    "geoutils.raster.core.Mask": "geoutils.Mask",
+    "geoutils.raster.raster.Raster": "geoutils.Raster",
+    "geoutils.raster.raster.Mask": "geoutils.Mask",
     "geoutils.raster.satimg.SatelliteImage": "geoutils.SatelliteImage",
     "geoutils.vector.Vector": "geoutils.Vector",
     "xdem.dem.DEM": "xdem.DEM",
